+++ src/main/java/ru/alfabouh/jgems3d/engine/audio/sound/SoundListener.java
@@ -1,7 +1,7 @@
 package ru.alfabouh.jgems3d.engine.audio.sound;
 
-import org.joml.Matrix4d;
-import org.joml.Vector3d;
+import org.joml.Matrix4f;
+import org.joml.Vector3f;
 import org.lwjgl.openal.AL10;
 import org.lwjgl.openal.ALC10;
 import org.lwjgl.system.MemoryUtil;
@@ -9,14 +9,14 @@ import ru.alfabouh.jgems3d.engine.JGems;
 import ru.alfabouh.jgems3d.engine.audio.SoundManager;
 
 public class SoundListener {
-    public static void updateOrientationAndPosition(Matrix4d cameraMatrix, Vector3d position) {
+    public static void updateOrientationAndPosition(Matrix4f cameraMatrix, Vector3f position) {
         if (ALC10.alcGetCurrentContext() == MemoryUtil.NULL) {
             return;
         }
         AL10.alListener3f(AL10.AL_POSITION, (float) position.x, (float) position.y, (float) position.z);
-        Vector3d at = new Vector3d();
+        Vector3f at = new Vector3f();
         cameraMatrix.positiveZ(at).negate();
-        Vector3d up = new Vector3d();
+        Vector3f up = new Vector3f();
         cameraMatrix.positiveY(up);
         float[] data = new float[]{(float) at.x, (float) at.y, (float) at.z, (float) up.x, (float) up.y, (float) up.z};
         AL10.alListenerfv(AL10.AL_ORIENTATION, data);
