+++ src/main/java/ru/alfabouh/jgems3d/engine/physics/triggers/Zone.java
@@ -1,21 +1,21 @@
 package ru.alfabouh.jgems3d.engine.physics.triggers;
 
-import org.joml.Vector3d;
+import org.joml.Vector3f;
 
 public class Zone {
-    private final Vector3d location;
-    private final Vector3d size;
+    private final Vector3f location;
+    private final Vector3f size;
 
-    public Zone(Vector3d location, Vector3d size) {
+    public Zone(Vector3f location, Vector3f size) {
         this.location = location;
         this.size = size;
     }
 
-    public Vector3d getLocation() {
-        return new Vector3d(this.location);
+    public Vector3f getLocation() {
+        return new Vector3f(this.location);
     }
 
-    public Vector3d getSize() {
-        return new Vector3d(this.size);
+    public Vector3f getSize() {
+        return new Vector3f(this.size);
     }
 }
