+++ src/main/java/ru/alfabouh/jgems3d/engine/render/opengl/scene/fabric/render/base/IRenderFabric.java
@@ -4,7 +4,7 @@ import ru.alfabouh.jgems3d.engine.render.opengl.scene.objects.IRenderObject;
 import ru.alfabouh.jgems3d.engine.render.opengl.scene.components.base.SceneRenderBase;
 
 public interface IRenderFabric {
-    void onRender(double partialTicks, SceneRenderBase sceneRenderBase, IRenderObject renderItem);
+    void onRender(float partialTicks, SceneRenderBase sceneRenderBase, IRenderObject renderItem);
 
     void onStartRender(IRenderObject renderItem);
