+++ src/main/java/ru/alfabouh/jgems3d/engine/render/opengl/scene/world/camera/ICamera.java
@@ -1,9 +1,9 @@
 package ru.alfabouh.jgems3d.engine.render.opengl.scene.world.camera;
 
-import org.joml.Vector3d;
+import org.joml.Vector3f;
 
 public interface ICamera {
-    Vector3d getCamPosition();
-    Vector3d getCamRotation();
-    void updateCamera(double partialTicks);
+    Vector3f getCamPosition();
+    Vector3f getCamRotation();
+    void updateCamera(float partialTicks);
 }
